requirements.txt
pystache==0.6.5 
static==1.1.1
blinker==1.8.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
filelock==3.14.0
Flask==3.0.3
fsspec==2024.5.0
gunicorn==22.0.0
huggingface-hub==0.23.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
packaging==24.0
pillow==10.3.0
PyYAML==6.0.1
regex==2024.5.15
requests==2.31.0
safetensors==0.4.3
sympy==1.12
tokenizers==0.19.1
torch==2.3.0
torchaudio==2.3.0
torchvision==0.18.0
tqdm==4.66.4
transformers==4.40.2
typing_extensions==4.11.0
urllib3==2.2.1
Werkzeug==3.0.3
